extract formatting staticmethods into separate files
