+++ isort/natural.py
@@ -26,7 +26,7 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 """
 import re
-from typing import Any, Callable, Iterable, List, Optional
+from typing import Any, Callable, Iterable, List, Mapping, Optional
 
 
 def _atoi(text: str) -> Any:
@@ -37,9 +37,46 @@ def _natural_keys(text: str) -> List[Any]:
     return [_atoi(c) for c in re.split(r'(\d+)', text)]
 
 
+def get_naturally_sortable_module_name(
+    module_name: str,
+    config: Mapping[str, Any],
+    sub_imports: bool = False,
+    ignore_case: bool = False,
+    section_name: Optional[Any] = None
+) -> str:
+    dots = 0
+    while module_name.startswith('.'):
+        dots += 1
+        module_name = module_name[1:]
+
+    if dots:
+        module_name = '{} {}'.format(('.' * dots), module_name)
+
+    prefix = ""
+    if ignore_case:
+        module_name = str(module_name).lower()
+    else:
+        module_name = str(module_name)
+
+    if sub_imports and config['order_by_type']:
+        if module_name.isupper() and len(module_name) > 1:
+            prefix = "A"
+        elif module_name[0:1].isupper():
+            prefix = "B"
+        else:
+            prefix = "C"
+    module_name = module_name.lower()
+    if section_name is None or 'length_sort_' + str(section_name).lower() not in config:
+        length_sort = config['length_sort']
+    else:
+        length_sort = config['length_sort_' + str(section_name).lower()]
+    return "{0}{1}{2}".format(module_name in config['force_to_top'] and "A" or "B", prefix,
+                              length_sort and (str(len(module_name)) + ":" + module_name) or module_name)
+
+
 def nsorted(
     to_sort: Iterable[str],
-    key: Optional[Callable[[str], Any]] = None
+    key: Optional[Callable[[str], str]] = None
 ) -> List[str]:
     """Returns a naturally sorted list"""
     if key is None:
